test: add integration test
Verify that we can use Daemon and DaemonConfig from outside, to prevent
bugs like 9bde11525.
